refactor(web): merge Copyable and CopyableLink components
